refactor(player): replace timer-based animation with event-driven approach
Replace useBeltProgressAnimation's useEffect+setTimeout orchestration
with onTransitionEnd and derived state. Extract BeltProgressHint into
its own belt-progress.tsx file for better separation of concerns.
